+++ CHANGELOG.md
@@ -8,10 +8,14 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+The release adds support for sending metrics directly to Splunk Observability Cloud.
+
 ### Add
 
 - Add the `WithIDGenerator` option to
   `github.com/signalfx/splunk-otel-go/distro`. (#2634)
+- Metrics are sent directly to Splunk Observability Cloud when `SPLUNK_REALM` is
+  set. (#2637)
 
 ## [1.10.0] - 2023-11-10